server/requirement.txt
Flask
pandas
pypdf
Flask-Cors
python-docx
nltk
num2words
scikit-learn
pymilvus
gensim